storage: factor out replica quota and destroy methods
Release note: None
