+++ pkg/storage/replica.go
@@ -81,11 +81,6 @@ const (
 	splitTxnName         = "split"
 	mergeTxnName         = "merge"
 
-	// MaxQuotaReplicaLivenessDuration is the maximum duration that a replica
-	// can remain inactive while still being counting against the range's
-	// available proposal quota.
-	MaxQuotaReplicaLivenessDuration = 10 * time.Second
-
 	defaultReplicaRaftMuWarnThreshold = 500 * time.Millisecond
 )
 
@@ -176,38 +171,6 @@ func (d *atomicDescString) store(replicaID roachpb.ReplicaID, desc *roachpb.Rang
 // using a lock, the copy might be inconsistent.
 func (d *atomicDescString) String() string {
 	return *(*string)(atomic.LoadPointer(&d.strPtr))
-}
-
-// DestroyReason indicates if a replica is alive, destroyed, corrupted or pending destruction.
-type DestroyReason int
-
-const (
-	// The replica is alive.
-	destroyReasonAlive DestroyReason = iota
-	// The replica has been corrupted.
-	destroyReasonCorrupted
-	// The replica has been marked for GC, but hasn't been GCed yet.
-	destroyReasonRemovalPending
-	// The replica has been GCed.
-	destroyReasonRemoved
-	// The replica has been merged into its left-hand neighbor, but its left-hand
-	// neighbor hasn't yet subsumed it.
-	destroyReasonMergePending
-)
-
-type destroyStatus struct {
-	reason DestroyReason
-	err    error
-}
-
-// IsAlive returns true when a replica is alive.
-func (s destroyStatus) IsAlive() bool {
-	return s.reason == destroyReasonAlive
-}
-
-// RemovedOrCorrupt returns true if a replica has either been removed or is corrupted.
-func (s destroyStatus) RemovedOrCorrupt() bool {
-	return (s.reason == destroyReasonCorrupted) || (s.reason == destroyReasonRemoved)
 }
 
 func (s *destroyStatus) Set(err error, reason DestroyReason) {
@@ -799,107 +762,6 @@ func (r *Replica) String() string {
 	return fmt.Sprintf("[n%d,s%d,r%s]", r.store.Ident.NodeID, r.store.Ident.StoreID, &r.rangeStr)
 }
 
-func (r *Replica) preDestroyRaftMuLocked(
-	ctx context.Context,
-	reader engine.Reader,
-	batch engine.Batch,
-	nextReplicaID roachpb.ReplicaID,
-	destroyData bool,
-) error {
-	desc := r.Desc()
-	err := clearRangeData(ctx, desc, reader, batch, destroyData)
-	if err != nil {
-		return err
-	}
-
-	// Save a tombstone to ensure that replica IDs never get reused.
-	//
-	// NB: Legacy tombstones (which are in the replicated key space) are wiped
-	// in clearRangeData, but that's OK since we're writing a new one in the same
-	// batch (and in particular, sequenced *after* the wipe).
-	return r.setTombstoneKey(ctx, batch, nextReplicaID)
-}
-
-func (r *Replica) postDestroyRaftMuLocked(ctx context.Context, ms enginepb.MVCCStats) error {
-	// Suggest the cleared range to the compactor queue.
-	//
-	// TODO(benesch): we would ideally atomically suggest the compaction with
-	// the deletion of the data itself.
-	desc := r.Desc()
-	r.store.compactor.Suggest(ctx, storagepb.SuggestedCompaction{
-		StartKey: roachpb.Key(desc.StartKey),
-		EndKey:   roachpb.Key(desc.EndKey),
-		Compaction: storagepb.Compaction{
-			Bytes:            ms.Total(),
-			SuggestedAtNanos: timeutil.Now().UnixNano(),
-		},
-	})
-
-	// NB: we need the nil check below because it's possible that we're GC'ing a
-	// Replica without a replicaID, in which case it does not have a sideloaded
-	// storage.
-	//
-	// TODO(tschottdorf): at node startup, we should remove all on-disk
-	// directories belonging to replicas which aren't present. A crash before a
-	// call to postDestroyRaftMuLocked will currently leave the files around
-	// forever.
-	if r.raftMu.sideloaded != nil {
-		return r.raftMu.sideloaded.Clear(ctx)
-	}
-	return nil
-}
-
-// destroyRaftMuLocked deletes data associated with a replica, leaving a
-// tombstone. If `destroyData` is true, data in all of the range's keyspaces
-// will be deleted. Otherwise, only data in the range-ID local keyspace will be
-// deleted. Requires that Replica.raftMu is held.
-func (r *Replica) destroyRaftMuLocked(ctx context.Context, nextReplicaID roachpb.ReplicaID) error {
-	startTime := timeutil.Now()
-
-	ms := r.GetMVCCStats()
-
-	const destroyData = true
-	batch := r.Engine().NewWriteOnlyBatch()
-	defer batch.Close()
-	if err := r.preDestroyRaftMuLocked(ctx, r.Engine(), batch, nextReplicaID, destroyData); err != nil {
-		return err
-	}
-	preTime := timeutil.Now()
-
-	// We need to sync here because we are potentially deleting sideloaded
-	// proposals from the file system next. We could write the tombstone only in
-	// a synchronous batch first and then delete the data alternatively, but
-	// then need to handle the case in which there is both the tombstone and
-	// leftover replica data.
-	if err := batch.Commit(true); err != nil {
-		return err
-	}
-	commitTime := timeutil.Now()
-
-	if err := r.postDestroyRaftMuLocked(ctx, ms); err != nil {
-		return err
-	}
-
-	log.Infof(ctx, "removed %d (%d+%d) keys in %0.0fms [clear=%0.0fms commit=%0.0fms]",
-		ms.KeyCount+ms.SysCount, ms.KeyCount, ms.SysCount,
-		commitTime.Sub(startTime).Seconds()*1000,
-		preTime.Sub(startTime).Seconds()*1000,
-		commitTime.Sub(preTime).Seconds()*1000)
-	return nil
-}
-
-func (r *Replica) cancelPendingCommandsLocked() {
-	r.mu.AssertHeld()
-	for _, p := range r.mu.proposals {
-		r.cleanupFailedProposalLocked(p)
-		// NB: each proposal needs its own version of the error (i.e. don't try to
-		// share the error across proposals).
-		p.finishApplication(proposalResult{
-			Err: roachpb.NewError(roachpb.NewAmbiguousResultError("removing replica")),
-		})
-	}
-}
-
 // cleanupFailedProposalLocked cleans up after a proposal that has failed. It
 // clears any references to the proposal and releases associated quota.
 func (r *Replica) cleanupFailedProposalLocked(p *ProposalData) {
@@ -916,34 +778,6 @@ func (r *Replica) cleanupFailedProposalLocked(p *ProposalData) {
 		r.mu.proposalQuota.add(int64(cmdSize))
 		delete(r.mu.commandSizes, p.idKey)
 	}
-}
-
-// setTombstoneKey writes a tombstone to disk to ensure that replica IDs never
-// get reused. It determines what the minimum next replica ID can be using
-// the provided nextReplicaID and the Replica's own ID.
-//
-// We have to be careful to set the right key, since a replica can be using an
-// ID that it hasn't yet received a RangeDescriptor for if it receives raft
-// requests for that replica ID (as seen in #14231).
-func (r *Replica) setTombstoneKey(
-	ctx context.Context, eng engine.ReadWriter, externalNextReplicaID roachpb.ReplicaID,
-) error {
-	r.mu.Lock()
-	nextReplicaID := r.mu.state.Desc.NextReplicaID
-	if nextReplicaID < externalNextReplicaID {
-		nextReplicaID = externalNextReplicaID
-	}
-	if nextReplicaID > r.mu.minReplicaID {
-		r.mu.minReplicaID = nextReplicaID
-	}
-	r.mu.Unlock()
-
-	tombstoneKey := keys.RaftTombstoneKey(r.RangeID)
-	tombstone := &roachpb.RaftTombstone{
-		NextReplicaID: nextReplicaID,
-	}
-	return engine.MVCCPutProto(ctx, eng, nil, tombstoneKey,
-		hlc.Timestamp{}, nil, tombstone)
 }
 
 func (r *Replica) setReplicaID(replicaID roachpb.ReplicaID) error {
@@ -1040,234 +874,6 @@ func (r *Replica) setReplicaIDRaftMuLockedMuLocked(replicaID roachpb.ReplicaID) 
 	}
 
 	return nil
-}
-
-func (r *Replica) maybeAcquireProposalQuota(ctx context.Context, quota int64) error {
-	r.mu.RLock()
-	quotaPool := r.mu.proposalQuota
-	desc := *r.mu.state.Desc
-	r.mu.RUnlock()
-
-	// Quota acquisition only takes place on the leader replica,
-	// r.mu.proposalQuota is set to nil if a node is a follower (see
-	// updateProposalQuotaRaftMuLocked). For the cases where the range lease
-	// holder is not the same as the range leader, i.e. the lease holder is a
-	// follower, r.mu.proposalQuota == nil. This means all quota acquisitions
-	// go through without any throttling whatsoever but given how short lived
-	// these scenarios are we don't try to remedy any further.
-	//
-	// NB: It is necessary to allow proposals with a nil quota pool to go
-	// through, for otherwise a follower could never request the lease.
-
-	if quotaPool == nil {
-		return nil
-	}
-
-	if !quotaPoolEnabledForRange(desc) {
-		return nil
-	}
-
-	// Trace if we're running low on available proposal quota; it might explain
-	// why we're taking so long.
-	if log.HasSpanOrEvent(ctx) {
-		if q := quotaPool.approximateQuota(); q < quotaPool.maxQuota()/10 {
-			log.Eventf(ctx, "quota running low, currently available ~%d", q)
-		}
-	}
-
-	return quotaPool.acquire(ctx, quota)
-}
-
-func quotaPoolEnabledForRange(desc roachpb.RangeDescriptor) bool {
-	// The NodeLiveness range does not use a quota pool. We don't want to
-	// throttle updates to the NodeLiveness range even if a follower is falling
-	// behind because this could result in cascading failures.
-	return !bytes.HasPrefix(desc.StartKey, keys.NodeLivenessPrefix)
-}
-
-func (r *Replica) updateProposalQuotaRaftMuLocked(
-	ctx context.Context, lastLeaderID roachpb.ReplicaID,
-) {
-	r.mu.Lock()
-	defer r.mu.Unlock()
-
-	if r.mu.replicaID == 0 {
-		// The replica was created from preemptive snapshot and has not been
-		// added to the Raft group.
-		return
-	}
-
-	// We need to check if the replica is being destroyed and if so, unblock
-	// all ongoing and subsequent quota acquisition goroutines (if any).
-	//
-	// TODO(irfansharif): There is still a potential problem here that leaves
-	// clients hanging if the replica gets destroyed but this code path is
-	// never taken. Moving quota pool draining to every point where a
-	// replica can get destroyed is an option, alternatively we can clear
-	// our leader status and close the proposalQuota whenever the replica is
-	// destroyed.
-	if r.mu.destroyStatus.RemovedOrCorrupt() {
-		if r.mu.proposalQuota != nil {
-			r.mu.proposalQuota.close()
-		}
-		r.mu.proposalQuota = nil
-		r.mu.lastUpdateTimes = nil
-		r.mu.quotaReleaseQueue = nil
-		r.mu.commandSizes = nil
-		return
-	}
-
-	if r.mu.leaderID != lastLeaderID {
-		if r.mu.replicaID == r.mu.leaderID {
-			// We're becoming the leader.
-			r.mu.proposalQuotaBaseIndex = r.mu.lastIndex
-
-			if r.mu.proposalQuota != nil {
-				log.Fatal(ctx, "proposalQuota was not nil before becoming the leader")
-			}
-			if releaseQueueLen := len(r.mu.quotaReleaseQueue); releaseQueueLen != 0 {
-				log.Fatalf(ctx, "len(r.mu.quotaReleaseQueue) = %d, expected 0", releaseQueueLen)
-			}
-			if commandSizesLen := len(r.mu.commandSizes); commandSizesLen != 0 {
-				log.Fatalf(ctx, "len(r.mu.commandSizes) = %d, expected 0", commandSizesLen)
-			}
-
-			// Raft may propose commands itself (specifically the empty
-			// commands when leadership changes), and these commands don't go
-			// through the code paths where we acquire quota from the pool. To
-			// offset this we reset the quota pool whenever leadership changes
-			// hands.
-			r.mu.proposalQuota = newQuotaPool(r.store.cfg.RaftProposalQuota)
-			r.mu.lastUpdateTimes = make(map[roachpb.ReplicaID]time.Time)
-			now := timeutil.Now()
-			for _, desc := range r.mu.state.Desc.Replicas {
-				r.mu.lastUpdateTimes.update(desc.ReplicaID, now)
-			}
-			r.mu.commandSizes = make(map[storagebase.CmdIDKey]int)
-		} else if r.mu.proposalQuota != nil {
-			// We're becoming a follower.
-
-			// We unblock all ongoing and subsequent quota acquisition
-			// goroutines (if any).
-			r.mu.proposalQuota.close()
-			r.mu.proposalQuota = nil
-			r.mu.lastUpdateTimes = nil
-			r.mu.quotaReleaseQueue = nil
-			r.mu.commandSizes = nil
-		}
-		return
-	} else if r.mu.proposalQuota == nil {
-		if r.mu.replicaID == r.mu.leaderID {
-			log.Fatal(ctx, "leader has uninitialized proposalQuota pool")
-		}
-		// We're a follower.
-		return
-	}
-
-	// We're still the leader.
-
-	// TODO(peter): Can we avoid retrieving the Raft status on every invocation
-	// in order to avoid the associated allocation? Tracking the progress
-	// ourselves via looking at MsgAppResp messages would be overkill. Perhaps
-	// another accessor on RawNode.
-	status := r.raftStatusRLocked()
-	if status == nil {
-		log.Fatal(ctx, "leader with nil RaftStatus")
-	}
-
-	// Find the minimum index that active followers have acknowledged.
-	now := timeutil.Now()
-	minIndex := status.Commit
-	for _, rep := range r.mu.state.Desc.Replicas {
-		// Only consider followers that that have "healthy" RPC connections.
-
-		if err := r.store.cfg.NodeDialer.ConnHealth(rep.NodeID); err != nil {
-			continue
-		}
-
-		// Only consider followers that are active.
-		if !r.mu.lastUpdateTimes.isFollowerActive(ctx, rep.ReplicaID, now) {
-			continue
-		}
-		if progress, ok := status.Progress[uint64(rep.ReplicaID)]; ok {
-			// Note that the Match field has different semantics depending on
-			// the State.
-			//
-			// In state ProgressStateReplicate, the Match index is optimistically
-			// updated whenever a message is *sent* (not received). Due to Raft
-			// flow control, only a reasonably small amount of data can be en
-			// route to a given follower at any point in time.
-			//
-			// In state ProgressStateProbe, the Match index equals Next-1, and
-			// it tells us the leader's optimistic best guess for the right log
-			// index (and will try once per heartbeat interval to update its
-			// estimate). In the usual case, the follower responds with a hint
-			// when it rejects the first probe and the leader replicates or
-			// sends a snapshot. In the case in which the follower does not
-			// respond, the leader reduces Match by one each heartbeat interval.
-			// But if the follower does not respond, we've already filtered it
-			// out above. We use the Match index as is, even though the follower
-			// likely isn't there yet because that index won't go up unless the
-			// follower is actually catching up, so it won't cause it to fall
-			// behind arbitrarily.
-			//
-			// Another interesting tidbit about this state is that the Paused
-			// field is usually true as it is used to limit the number of probes
-			// (i.e. appends) sent to this follower to one per heartbeat
-			// interval.
-			//
-			// In state ProgressStateSnapshot, the Match index is the last known
-			// (possibly optimistic, depending on previous state) index before
-			// the snapshot went out. Once the snapshot applies, the follower
-			// will enter ProgressStateReplicate again. So here the Match index
-			// works as advertised too.
-
-			// Only consider followers who are in advance of the quota base
-			// index. This prevents a follower from coming back online and
-			// preventing throughput to the range until it has caught up.
-			if progress.Match < r.mu.proposalQuotaBaseIndex {
-				continue
-			}
-			if progress.Match > 0 && progress.Match < minIndex {
-				minIndex = progress.Match
-			}
-			// If this is the most recently added replica and it has caught up, clear
-			// our state that was tracking it. This is unrelated to managing proposal
-			// quota, but this is a convenient place to do so.
-			if rep.ReplicaID == r.mu.lastReplicaAdded && progress.Match >= status.Commit {
-				r.mu.lastReplicaAdded = 0
-				r.mu.lastReplicaAddedTime = time.Time{}
-			}
-		}
-	}
-
-	if r.mu.proposalQuotaBaseIndex < minIndex {
-		// We've persisted minIndex - r.mu.proposalQuotaBaseIndex entries to
-		// the raft log on all 'active' replicas since last we checked,
-		// we 'should' be able to release the difference back to
-		// the quota pool. But consider the scenario where we have a single
-		// replica that we're writing to, we only construct the
-		// quotaReleaseQueue when entries 'come out' of Raft via
-		// raft.Ready.CommittedEntries. The minIndex computed above uses the
-		// replica's commit index which is independent of whether or we've
-		// iterated over the entirety of raft.Ready.CommittedEntries and
-		// therefore may not have all minIndex - r.mu.proposalQuotaBaseIndex
-		// command sizes in our quotaReleaseQueue.  Hence we only process
-		// min(minIndex - r.mu.proposalQuotaBaseIndex, len(r.mu.quotaReleaseQueue))
-		// quota releases.
-		numReleases := minIndex - r.mu.proposalQuotaBaseIndex
-		if qLen := uint64(len(r.mu.quotaReleaseQueue)); qLen < numReleases {
-			numReleases = qLen
-		}
-		sum := 0
-		for _, rel := range r.mu.quotaReleaseQueue[:numReleases] {
-			sum += rel
-		}
-		r.mu.proposalQuotaBaseIndex += numReleases
-		r.mu.quotaReleaseQueue = r.mu.quotaReleaseQueue[numReleases:]
-
-		r.mu.proposalQuota.add(int64(sum))
-	}
 }
 
 // GetMinBytes gets the replica's minimum byte threshold.
